pyproject.toml
[project]
name = "shopscale"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faker>=39.0.0",
    "fastapi[standard]>=0.127.1",
    "httpx>=0.28.1",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.45",
]
